feat(host)!: supervise and restart providers on exit
